+++ lib/views/view_payees/view_payees.dart
@@ -9,11 +9,11 @@ import 'package:money/models/fields/fields.dart';
 import 'package:money/models/money_objects/payees/payee.dart';
 import 'package:money/models/money_objects/transactions/transactions.dart';
 import 'package:money/storage/data/data.dart';
+import 'package:money/views/adaptive_view/adaptive_list/transactions/list_view_transactions.dart';
 import 'package:money/views/view_money_objects.dart';
 import 'package:money/views/view_payees/merge_payees.dart';
 import 'package:money/widgets/center_message.dart';
 import 'package:money/widgets/chart.dart';
-import 'package:money/views/adaptive_view/adaptive_list/transactions/list_view_transactions.dart';
 import 'package:money/widgets/date_range_time_line.dart';
 import 'package:money/widgets/gaps.dart';
 import 'package:money/widgets/mini_timeline_daily.dart';
@@ -81,13 +81,6 @@ class ViewPayeesState extends ViewForMoneyObjectsState {
     required final bool showAsNativeCurrency,
   }) {
     return _getSubViewContentForTransactions(selectedIds);
-  }
-
-  @override
-  void onDeleteConfirmedByUser(final MoneyObject instance) {
-    setState(() {
-      Data().payees.deleteItem(instance);
-    });
   }
 
   @override